requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.24
SQLAlchemy==2.0.43
python-dotenv==1.0.1
httpx==0.27.2
mariadb==1.1.13
